Integrate with previous project

+++ login.php
@@ -12,11 +12,14 @@ include("functions.php");
 if (isset($_GET["getUserAction"])) {
 	$username = $_GET["username"];
 	
-	if (isUsernameExists($username)) {
-		// istnieje
-		$userID = getUserIDFromDB($username);
-		$mask = getMask($userID);
-?>		
+	$registered = "users";
+	$unregistered = "unregistered_users";
+
+	if (isUsernameExistsInTable($username, $registered) && !isUserBlocked($username, $registered)) {
+		if (!isEnableToLockUser($username, $registered) && !isEnableToBlockUser($username, $registered)) {
+			$userID = getUserIDFromDB($username);
+			$mask = getMask($userID);
+			?>		
 		<form id="login" action="login.php" method="get">
 	Hasło:<br />
 <?php
@@ -34,9 +37,83 @@ if (isset($_GET["getUserAction"])) {
 			<input type="submit" name="loginAction" value="Zaloguj" />
 		</form>
 <?php
+		}
+		
+		if (isEnableToLockUser($username, $registered) && !isEnableToBlockUser($username, $registered)) {
+			echo "Zablokowano możliwość logowania użytkownika $username na " . timeToEndLock($username, $registered) . " sekund.<br /><br />";
+			echo "<a href='index.php'>Powrot</a>";
+		}
+		
+		if (isEnableToBlockUser($username, $registered)) {
+			blockUser($username, $registered);
+			echo "Konto zablokowane.<br />Odpowiedz na wcześniej zdefiniowane pytanie, aby odblokować dostęp do konta. <br /><br />";
+			include("retrieve.php");
+		}
+	} else if (isUserBlocked($username, $registered)) {
+		echo "Konto zablokowane.<br />Odpowiedz na wcześniej zdefiniowane pytanie, aby odblokować dostęp do konta. <br /><br />";
+		include("retrieve.php");
+	} else if (isUserBlocked($username, $unregistered)) {
+		echo "Konto zablokowane.<br />Odpowiedz na wcześniej zdefiniowane pytanie, aby odblokować dostęp do konta. <br /><br />";
+		include("retrieve.php");
 	} else {
-		// nie istnieje
-		echo "! user nie istnieje";
+		if (isUsernameExistsInTable($username, $unregistered)) {
+			if (!isEnableToLockUser($username, $unregistered) && !isEnableToBlockUser($username, $registered)) {
+				$userID = getUserIDFromDBunreg($username);
+				$mask = getMask($userID);
+?>		
+				<form id="login" action="login.php" method="get">
+					Hasło:<br />
+<?php
+					$disabled = "";
+				for ($i = 0; $i < count($mask); $i++) {
+					if ($mask[$i] == '1') {
+						$disabled = "";
+					} else {
+						$disabled = "disabled";
+					}
+					echo "<input type='text' name='p[$i]' maxlength='1' $disabled size='1' /> ";
+				}
+?>
+					<input type="text" name="username" value="<?php echo $username; ?>" hidden />
+					<input type="submit" name="loginAction" value="Zaloguj" />
+				</form>
+<?php
+			}
+			
+			if (isEnableToLockUser($username, $unregistered) && !isEnableToBlockUser($username, $unregistered)) {
+				echo "Zablokowano możliwość logowania użytkownika $username na " . timeToEndLock($username, $unregistered) . " sekund.<br /><br />";
+				echo "<a href='index.php'>Powrot</a>";
+			}
+			
+			if (isEnableToBlockUser($username, $unregistered)) {
+				blockUser($username, $unregistered);
+				echo "Konto zablokowane.<br />Odpowiedz na wcześniej zdefiniowane pytanie, aby odblokować dostęp do konta. <br /><br />";
+				include("retrieve.php");
+			}
+		} else {
+			addUnregisteredLoginAttempt($username);
+			
+			$userID = getUserIDFromDBunreg($username);
+				$mask = getMask($userID);
+?>		
+				<form id="login" action="login.php" method="get">
+					Hasło:<br />
+<?php
+					$disabled = "";
+				for ($i = 0; $i < count($mask); $i++) {
+					if ($mask[$i] == '1') {
+						$disabled = "";
+					} else {
+						$disabled = "disabled";
+					}
+					echo "<input type='text' name='p[$i]' maxlength='1' $disabled size='1' /> ";
+				}
+?>
+					<input type="text" name="username" value="<?php echo $username; ?>" hidden />
+					<input type="submit" name="loginAction" value="Zaloguj" />
+				</form>
+<?php
+		}
 	}
 }
 ?>
@@ -50,23 +127,50 @@ if (isset($_GET["loginAction"])) {
  	$partialPassword = $_GET["p"];
 	$username = $_GET["username"];
 	
+	$registered = "users";
+	$unregistered = "unregistered_users";
+	
 	$mask = retrieveMask($partialPassword, $username);
 	
-	if ($mask != null) {
+	if ($mask != null && isUsernameExistsInTable($username, $registered)) {
 		$partialHash = getPartialHash($username, $mask);
 		$salt = getUserSaltFromDB($username);
 		$currentHash = reCreatePartialPasswordHash(str_split($mask), implode($partialPassword), $salt);
-				
+		
 		if ($partialHash == $currentHash) {
 			echo "Poprawne haslo";
 			setPasswordChecked($mask, $username);
-			// ustawic last_used = 0 i is_used = 1
+			login($username);
+			header("location: messages.php");
 		} else {
-			echo "Niepoprawne haslo";
+			increaseUserLoginAttempts($username, $registered);
+			echo "Nieprawidłowe hasło.<br /><br />";
 			echo "<a href='index.php'>Powrot</a>";
 		}
+	} else if (isUsernameExistsInTable($username, $registered)) {
+		increaseUserLoginAttempts($username, $registered);
+		echo "Nieprawidłowe hasło.<br /><br />";
+		echo "<a href='index.php'>Powrot</a>";
+	} else if (isUsernameExistsInTable($username, $unregistered)) {
+		increaseUserLoginAttempts($username, $unregistered);
+		echo "Nieprawidłowe hasło.<br /><br />";
+		echo "<a href='index.php'>Powrot</a>";
+	}
+}
+
+if (isset($_GET["unBlockAction"])) {
+	$username = trim($_GET["username"]);
+	$answer = trim($_GET["answer"]);
+	
+	if (!isAnswerMatch($username, $answer)) {
+		echo "Odpowiedź nieprawidłowa. Spróbuj ponownie.<br /><br />";
+		include("retrieve.php");
 	} else {
-		echo "Niepoprawne haslo";
+		$registered = "users";
+		$unregistered = "unregistered_users";
+		unBlockUser($username, $registered);
+		unBlockUser($username, $unregistered);
+		echo "Konto odblokowane<br /><br />";
 		echo "<a href='index.php'>Powrot</a>";
 	}
 }
